refactor(gibs): share url builders via proxy-constants
