+++ packages/proxy-constants/tsconfig.json
@@ -13,6 +13,6 @@
     "outDir": "./dist",
     "rootDir": "./"
   },
-  "include": ["index.js"],
+  "include": ["index.ts"],
   "exclude": ["node_modules", "dist"]
 }
